requirements.txt
streamlit~=1.33.0
openai~=1.16.2
boto3~=1.34.79
docx2txt~=0.8
python-dotenv~=1.0.1
PyPDF2~=3.0.1
pillow~=10.2.0
pymongo~=4.6.3
langchain~=0.1.14